+++ src/css/railyardShared.module.css
@@ -1,0 +1,65 @@
+.select {
+    appearance: none;
+    border: 1px solid color-mix(in srgb, var(--ifm-border-color) 50%, #10b981 50%);
+    border-radius: 12px;
+    padding: 0.5rem 2.2rem 0.5rem 0.7rem;
+    font-size: 0.88rem;
+    font-weight: 600;
+    color: var(--ifm-font-color-base);
+    background:
+            linear-gradient(45deg, transparent 50%, #10b981 50%),
+            linear-gradient(135deg, #10b981 50%, transparent 50%),
+            linear-gradient(
+                    180deg,
+                    color-mix(in srgb, var(--ifm-background-color) 92%, #10b981 8%),
+                    var(--ifm-background-color)
+            );
+    background-position:
+            calc(100% - 16px) calc(50% - 2px),
+            calc(100% - 10px) calc(50% - 2px),
+            0 0;
+    background-size: 6px 6px, 6px 6px, 100% 100%;
+    background-repeat: no-repeat;
+    box-shadow: 0 2px 12px rgba(16, 185, 129, 0.08);
+    transition: border-color 0.16s ease, box-shadow 0.16s ease;
+}
+
+.select:focus {
+    outline: none;
+    border-color: #10b981;
+    box-shadow: 0 0 0 3px color-mix(in srgb, #10b981 24%, transparent);
+}
+
+.pagination {
+    border: 1px solid color-mix(in srgb, var(--ifm-border-color) 45%, #10b981 55%);
+    border-radius: 999px;
+    padding: 0.3rem;
+    background: linear-gradient(
+            145deg,
+            color-mix(in srgb, var(--ifm-alt-background-color) 74%, #10b981 26%),
+            color-mix(in srgb, var(--ifm-alt-background-color) 90%, #0039a6 10%)
+    );
+    box-shadow: 0 10px 24px rgba(0, 0, 0, 0.12);
+}
+
+.pagination button {
+    border: 1px solid color-mix(in srgb, var(--ifm-border-color) 60%, #10b981 40%);
+    border-radius: 999px;
+    padding: 0.3rem 0.78rem;
+    font-size: 0.76rem;
+    font-weight: 800;
+    letter-spacing: 0.02em;
+    background: color-mix(in srgb, var(--ifm-background-color) 90%, #10b981 10%);
+    transition: transform 0.14s ease, box-shadow 0.14s ease, border-color 0.14s ease;
+}
+
+.pagination button:hover:not(:disabled) {
+    transform: translateY(-1px);
+    border-color: #10b981;
+    box-shadow: 0 4px 10px rgba(16, 185, 129, 0.2);
+}
+
+.pagination span {
+    min-width: 7.5rem;
+    text-align: center;
+}
