+++ src/pages/railyard/maps.js
@@ -3,6 +3,16 @@ import Layout from "@theme/Layout";
 import Link from "@docusaurus/Link";
 import { translate } from "@docusaurus/Translate";
 import styles from "../../css/railyardMaps.module.css";
+import sharedStyles from "../../css/railyardShared.module.css";
+import {
+  compareValues,
+  flattenRecord,
+  formatTagLabel,
+  getFirstValue,
+  getPopulation,
+  getTitle,
+  normalizeImageList,
+} from "../../helpers/railyardHelpers";
 
 const SOURCE = {
   type: "maps",
@@ -10,135 +20,31 @@ const SOURCE = {
     "https://raw.githubusercontent.com/Subway-Builder-Modded/The-Railyard/main/maps/index.json",
 };
 
-const PAGE_SIZES = [9, 27, 54];
+const PAGE_SIZES = [12, 24, 48];
 
-const fieldPathLookup = {
-  title: ["name", "title", "displayName"],
-  description: ["description", "summary"],
-  author: ["author", "creator", "publisher"],
-  tags: ["tags", "categories", "labels"],
-  population: ["population", "cityPopulation", "populationEstimate"],
-  images: ["images", "gallery", "screenshots", "thumbnails"],
+const SORT_CONFIG = {
+  name: {
+    value: (item) => item.title,
+    defaultDirection: "asc",
+  },
+  population: {
+    value: (item) => item.population,
+    defaultDirection: "desc",
+  },
+  id: {
+    value: (item) => item.id,
+    defaultDirection: "asc",
+  },
 };
 
-function getFirstValue(record, key) {
-  for (const path of fieldPathLookup[key] || []) {
-    if (Object.prototype.hasOwnProperty.call(record, path) && record[path] != null) {
-      return record[path];
-    }
-  }
-  return null;
-}
-
-function flattenRecord(record, prefix = "") {
-  if (record == null || typeof record !== "object") {
-    return [{ key: prefix || "value", value: String(record) }];
-  }
-
-  const rows = [];
-  Object.entries(record).forEach(([key, value]) => {
-    const path = prefix ? `${prefix}.${key}` : key;
-    if (Array.isArray(value)) {
-      rows.push({ key: path, value: JSON.stringify(value) });
-    } else if (value && typeof value === "object") {
-      rows.push(...flattenRecord(value, path));
-    } else {
-      rows.push({ key: path, value: String(value) });
-    }
-  });
-  return rows;
-}
-
-function toTitleCaseTag(tag) {
-  return tag
-    .split("-")
-    .filter(Boolean)
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1).toLowerCase())
-    .join(" ");
-}
-
-function buildGithubImageCandidates(url) {
-  if (!url) return [];
-
-  const normalized = url.replace("/refs/heads/", "/").replace("?raw=true", "");
-  const candidates = [encodeURI(normalized)];
-
-  const githubMatch = normalized.match(
-    /^https:\/\/github\.com\/([^/]+)\/([^/]+)\/(?:blob|raw)\/([^/]+)\/(.+)$/,
-  );
-  if (githubMatch) {
-    const [, owner, repo, branch, rawPath] = githubMatch;
-    candidates.push(
-      `https://raw.githubusercontent.com/${owner}/${repo}/${branch}/${encodeURI(rawPath)}`,
-    );
-    candidates.push(`https://cdn.jsdelivr.net/gh/${owner}/${repo}@${branch}/${encodeURI(rawPath)}`);
-  }
-
-  const rawMatch = normalized.match(
-    /^https:\/\/raw\.githubusercontent\.com\/([^/]+)\/([^/]+)\/([^/]+)\/(.+)$/,
-  );
-  if (rawMatch) {
-    const [, owner, repo, branch, rawPath] = rawMatch;
-    candidates.push(`https://cdn.jsdelivr.net/gh/${owner}/${repo}@${branch}/${encodeURI(rawPath)}`);
-  }
-
-  return [...new Set(candidates)];
-}
-
-function normalizeImageList(manifest, id) {
-  const imageCandidates = getFirstValue(manifest, "images");
-  if (!imageCandidates) return [];
-
-  const toUrl = (value) => {
-    if (!value) return [];
-
-    if (typeof value === "string") {
-      if (value.startsWith("http")) return buildGithubImageCandidates(value);
-      const cleaned = value
-        .replace(/^\.\//, "")
-        .replace(new RegExp(`^maps/${id}/gallery/`), "")
-        .replace(new RegExp(`^${id}/gallery/`), "")
-        .replace(/^gallery\//, "")
-        .replace(/^maps\//, "")
-        .replace(/^\//, "");
-      const relative = `maps/${id}/gallery/${cleaned}`;
-      return [
-        `https://raw.githubusercontent.com/Subway-Builder-Modded/The-Railyard/main/${encodeURI(relative)}`,
-        `https://cdn.jsdelivr.net/gh/Subway-Builder-Modded/The-Railyard@main/${encodeURI(relative)}`,
-      ];
-    }
-
-    if (typeof value === "object") {
-      const nested = value.file || value.src || value.path || value.url;
-      return toUrl(nested);
-    }
-
-    return [];
+function getSortValue(sortBy) {
+  const [column, direction] = sortBy.split("-");
+  const config = SORT_CONFIG[column] || SORT_CONFIG.name;
+  const sortDirection = direction || config.defaultDirection;
+  return {
+    getValue: config.value,
+    descending: sortDirection === "desc",
   };
-
-  if (Array.isArray(imageCandidates)) {
-    return imageCandidates.map(toUrl).filter((candidateList) => candidateList.length > 0);
-  }
-
-  const singleUrl = toUrl(imageCandidates);
-  return singleUrl.length > 0 ? [singleUrl] : [];
-}
-
-function getTitle(manifest, id) {
-  return getFirstValue(manifest, "title") || id;
-}
-
-function getPopulation(manifest) {
-  const value = getFirstValue(manifest, "population");
-  const numeric = Number(value);
-  return Number.isFinite(numeric) ? numeric : null;
-}
-
-function compareValues(a, b) {
-  if (a == null && b == null) return 0;
-  if (a == null) return 1;
-  if (b == null) return -1;
-  return String(a).localeCompare(String(b), undefined, { numeric: true, sensitivity: "base" });
 }
 
 function MapPinPlaceholder() {
@@ -190,7 +96,7 @@ export default function RailyardMapsPage() {
   const [query, setQuery] = useState("");
   const [selectedTags, setSelectedTags] = useState([]);
   const [sortBy, setSortBy] = useState("name-asc");
-  const [pageSize, setPageSize] = useState(27);
+  const [pageSize, setPageSize] = useState(24);
   const [page, setPage] = useState(1);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState("");
@@ -248,7 +154,7 @@ export default function RailyardMapsPage() {
                     id: "railyard.maps.unknownAuthor",
                     message: "Unknown",
                   }),
-                images: normalizeImageList(manifest, id),
+                images: normalizeImageList(manifest, SOURCE.type, id),
                 fields: flattenRecord(manifest),
               };
             } catch {
@@ -298,10 +204,10 @@ export default function RailyardMapsPage() {
         item.title,
         item.description,
         item.author,
-        item.tags.join(" "),
-        item.fields.map((field) => `${field.key} ${field.value}`).join(" "),
+        item.tags.join("-"),
+        item.fields.map((field) => `${field.key} ${field.value}`).join("-"),
       ]
-        .join(" ")
+        .join("-")
         .toLowerCase();
       return haystack.includes(lowered);
     });
@@ -311,21 +217,9 @@ export default function RailyardMapsPage() {
     );
 
     return tagFiltered.sort((a, b) => {
-      switch (sortBy) {
-        case "name-desc":
-          return compareValues(b.title, a.title);
-        case "population-desc":
-          return compareValues(b.population, a.population);
-        case "population-asc":
-          return compareValues(a.population, b.population);
-        case "id-asc":
-          return compareValues(a.id, b.id);
-        case "id-desc":
-          return compareValues(b.id, a.id);
-        case "name-asc":
-        default:
-          return compareValues(a.title, b.title);
-      }
+      const { getValue, descending } = getSortValue(sortBy);
+      const comparison = compareValues(getValue(a), getValue(b));
+      return descending ? -comparison : comparison;
     });
   }, [items, query, selectedTags, sortBy]);
 
@@ -366,7 +260,9 @@ export default function RailyardMapsPage() {
     if (totalPages <= 1) return null;
 
     return (
-      <nav className={`${styles.pagination} ${isTop ? styles.paginationTop : ""}`}>
+      <nav
+        className={`${styles.pagination} ${sharedStyles.pagination} ${isTop ? styles.paginationTop : ""}`}
+      >
         <button
           type="button"
           disabled={safePage <= 1}
@@ -436,7 +332,11 @@ export default function RailyardMapsPage() {
           <div className={styles.controlRow}>
             <label>
               {translate({ id: "railyard.maps.sortBy", message: "Sort by" })}
-              <select value={sortBy} onChange={(event) => setSortBy(event.target.value)}>
+              <select
+                className={sharedStyles.select}
+                value={sortBy}
+                onChange={(event) => setSortBy(event.target.value)}
+              >
                 <option value="name-asc">
                   {translate({ id: "railyard.maps.sort.nameAsc", message: "Name (A → Z)" })}
                 </option>
@@ -467,6 +367,7 @@ export default function RailyardMapsPage() {
             <label>
               {translate({ id: "railyard.maps.cardsPerPage", message: "Cards per page" })}
               <select
+                className={sharedStyles.select}
                 value={pageSize}
                 onChange={(event) => setPageSize(Number(event.target.value))}
               >
@@ -489,7 +390,7 @@ export default function RailyardMapsPage() {
                   className={`${styles.tagButton} ${active ? styles.tagButtonActive : ""}`}
                   onClick={() => toggleTag(tag)}
                 >
-                  {toTitleCaseTag(tag)}
+                  {formatTagLabel(tag)}
                 </button>
               );
             })}
@@ -585,7 +486,7 @@ export default function RailyardMapsPage() {
                         toggleTag(tag);
                       }}
                     >
-                      {toTitleCaseTag(tag)}
+                      {formatTagLabel(tag)}
                     </button>
                   ))}
                 </div>
